+++ tests/test_om_prior/test_005_output_domain_file.csv
@@ -1,0 +1,2 @@
+,LAT,LON,LANDMASK,LATD,LOND,OCH4_AGRICULTURE,OCH4_LULUCF,OCH4_WASTE,OCH4_LIVESTOCK,OCH4_INDUSTRIAL,OCH4_STATIONARY,OCH4_TRANSPORT,OCH4_ELECTRICITY,OCH4_FUGITIVE,OCH4_TERMITE,OCH4_FIRE,OCH4_WETLANDS,OCH4_TOTAL
+0,-26.983160018920898,133.302001953125,0.39128163456916809,-26.980266571044922,133.302001953125,3.8221794892533283e-13,8.2839841777071689e-13,7.6806688034203811e-13,3.3252710482543991e-12,4.6406982245152562e-15,8.5852921845127254e-14,1.8562792898061025e-14,2.3204437472569127e-14,1.9068246493083643e-12,8.1061502996138124e-13,2.6130391795643748e-13,1.8596938045956645e-11,2.7011896889060713e-11
